Add files via upload
In this project, I scraped a novel from an HTML file sourced from the Project Gutenberg website, which hosts a vast collection of books. Using the Python requests package, I fetched the HTML content and then utilized the BeautifulSoup library to extract the text. Afterward, I employed the Natural Language ToolKit (nltk) to analyze the distribution of words. This process is a common NLP task that provides valuable insights into the text and serves as a foundation for more advanced textual data analysis.